requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv
requests

numpy<2
scikit-learn==1.4.2

torch==2.1.2+cpu
sentence-transformers==2.6.1
transformers==4.39.3
